Generator 1

+++ Assets/Scripts/Resources/ResourceBank.cs
@@ -3,7 +3,7 @@ public class ResourceBank
     private long _money;
     public long Money { get => _money; set => _money = value; }
 
-    public ResourceBank(int money)
+    public ResourceBank(long money)
     {
         Money = money;
     }
